회원 페이지 모두 완성. 이제 세션 추가 및 연결 필요

+++ src/main/java/com/example/prj2/board/controller/BoardController.java
@@ -4,6 +4,8 @@ package com.example.prj2.board.controller;
 import com.example.prj2.board.dto.BoardDto;
 import com.example.prj2.board.entity.Board;
 import com.example.prj2.board.service.BoardService;
+import com.example.prj2.member.Member;
+import jakarta.servlet.http.HttpSession;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -22,14 +24,23 @@ public class BoardController {
     private final BoardService boardService;
 
     @GetMapping("write")
-    public String write() {
-        return "board/write";
+    public String write(HttpSession session, RedirectAttributes rttr) {
+        Object login = session.getAttribute("login");
+        if (login == null) {
+            rttr.addFlashAttribute("loginRequired", "로그인이 필요합니다.");
+            return "redirect:/member/login";
+        } else {
+            return "board/write";
+
+        }
     }
 
     @PostMapping("write")
-    public String write(BoardDto boardDto, Model model) {
+    public String write(BoardDto boardDto, Model model, HttpSession session) {
         boolean result = boardService.write(boardDto);
         if (result) {
+            Member login = (Member) session.getAttribute("login");
+            boardDto.setAuthor(login.getId());
             return "redirect:/board/list";
         } else {
             model.addAttribute("error", "모든 항목을 입력해주세요");
@@ -39,9 +50,7 @@ public class BoardController {
     }
 
     @GetMapping("list")
-    public String list(@RequestParam(defaultValue = "1")
-                       Integer page,
-                       Model model) {
+    public String list(@RequestParam(defaultValue = "1") Integer page, Model model) {
         if (page == null || page < 1) page = 1;
 
         // 조회
@@ -52,10 +61,10 @@ public class BoardController {
     }
 
     @GetMapping("view")
-    public String view(String id, Model model,
-                       @RequestParam(required = false, defaultValue = "1")
-                       Integer page) {
-        var view = boardService.get(id);
+    public String view(String id, Model model, @RequestParam(required = false, defaultValue = "1") Integer page, HttpSession session) {
+        Member login = (Member) session.getAttribute("login");
+
+        var view = boardService.get(id, login);
 
         model.addAttribute("view", view);
         model.addAttribute("page", page);
@@ -63,16 +72,24 @@ public class BoardController {
     }
 
     @GetMapping("edit")
-    public String edit(String id, Model model) {
-        com.example.prj2.board.dto.BoardDto boardDto = boardService.get(id);
-        model.addAttribute("board", boardDto);
-        return "board/edit";
+    public String edit(String id, Model model, RedirectAttributes rttr, HttpSession session) {
+        Member login = (Member) session.getAttribute("login");
+        if (login == null) {
+            rttr.addFlashAttribute("loginRequired", "로그인이 필요합니다.");
+            return "redirect:/member/login";
+        } else {
+            BoardDto dto = boardService.get(id, login);
+            if (dto == null) {
+                rttr.addFlashAttribute("error", "작성자만 수정 할 수 있습니다.");
+                return "redirect:/board/list";
+            }
+            model.addAttribute("board", dto);
+            return "board/edit";
+        }
     }
 
     @PostMapping("edit")
-    public String edit(String id,
-                       com.example.prj2.board.dto.BoardDto boardDto,
-                       RedirectAttributes rttr) {
+    public String edit(String id, BoardDto boardDto, RedirectAttributes rttr) {
         boardService.update(boardDto);
         rttr.addFlashAttribute("update", id + "번 게시글이 수정되었습니다.");
 
@@ -80,9 +97,16 @@ public class BoardController {
     }
 
     @PostMapping("delete")
-    public String delete(String id, RedirectAttributes rttr) {
-        boardService.delete(id);
-        rttr.addFlashAttribute("deleteMessage", id + "번 게시물이 삭제되었습니다.");
-        return "redirect:/board/list";
+    public String delete(String id, RedirectAttributes rttr, HttpSession session) {
+        Object login = session.getAttribute("login");
+        if (login == null) {
+            rttr.addFlashAttribute("loginRequired", "로그인이 필요합니다.");
+            return "redirect:/member/login";
+        } else {
+            boardService.delete(id);
+            rttr.addFlashAttribute("deleteMessage", id + "번 게시물이 삭제되었습니다.");
+            return "redirect:/board/list";
+        }
     }
+
 }
